update/ rename GroupList Btn
